+++ akita/src/api/store.ts
@@ -1,14 +1,14 @@
 import { HashMap, ID } from './types';
 import { BehaviorSubject, Observable, ReplaySubject } from 'rxjs';
 import { distinctUntilChanged, map } from 'rxjs/operators';
-import { AkitaImmutabilityError, assertDecorator } from '../internal/error';
+import { AkitaError, AkitaImmutabilityError, assertDecorator } from '../internal/error';
 import { commit, isTransactionInProcess } from '../internal/transaction.internal';
 import { isFunction, isPlainObject } from '../internal/utils';
 import { deepFreeze } from '../internal/deep-freeze';
 import { configKey, StoreConfigOptions } from './store-config';
 import { __globalState } from '../internal/global-state';
+import { getAkitaConfig } from './config';
 
-/** Whether we are in dev mode */
 let __DEV__ = true;
 
 export const __stores__: { [storeName: string]: Store<any> } = {};
@@ -62,6 +62,8 @@ export class Store<S> {
 
   private _isPristine = true;
 
+  private _initialState: S;
+
   /**
    *
    * Initial the store with the state
@@ -75,6 +77,9 @@ export class Store<S> {
       payload: { store: this }
     });
     isDev() && assertDecorator(this.storeName, this.constructor.name);
+    if (getAkitaConfig().resettable) {
+      this._initialState = initialState;
+    }
   }
 
   setLoading(loading = false) {
@@ -152,6 +157,19 @@ export class Store<S> {
     }
 
     this.dispatch(this.storeValue, _rootDispatcher);
+  }
+
+  /**
+   * Resets the store to it's initial state and set the store to a pristine state.
+   */
+  reset() {
+    if (getAkitaConfig().resettable) {
+      __globalState.setAction({ type: 'Reset Store' });
+      this.setState(() => Object.assign({}, this._initialState));
+      this.setPristine();
+    } else {
+      throw new AkitaError(`You need to enable the reset functionality`);
+    }
   }
 
   /**