+++ akita/src/internal/crud.ts
@@ -1,9 +1,9 @@
-import { Entities, EntityState, HashMap, ID, Newable } from '../api/types';
+import { Entities, EntityState, HashMap, ID, Newable, AddOptions } from '../api/types';
 import { AkitaUpdateIdKeyError, assertEntityExists, assertEntityState } from './error';
 import { entityExists, isFunction, isPlainObject, resetActive } from './utils';
 
 export class CRUD {
-  _set<S, E>(state: S, entities: E[] | HashMap<E> | Entities<E>, entityClass: Newable<E>, idKey): S {
+  _set<S, E>(state: S, entities: E[] | HashMap<E> | Entities<E>, entityClass: Newable<E> | undefined, idKey): S {
     let ids, normalized;
 
     if ((entities as Entities<E>).ids && (entities as Entities<E>).entities) {
@@ -46,7 +46,7 @@ export class CRUD {
     };
   }
 
-  _add<S extends EntityState, E>(state: S, entities: E[], idKey): S {
+  _add<S extends EntityState, E>(state: S, entities: E[], idKey, options: AddOptions = {}): S {
     let addedEntities = {};
     let addedIds = [];
 
@@ -56,7 +56,8 @@ export class CRUD {
 
       if (!entityExists(entityId, state.entities)) {
         addedEntities[entityId] = entity;
-        addedIds.push(entityId);
+        if (options.prepend) addedIds.unshift(entityId);
+        else addedIds.push(entityId);
       }
     }
 
@@ -66,7 +67,7 @@ export class CRUD {
         ...state.entities,
         ...addedEntities
       },
-      ids: [...state.ids, ...addedIds]
+      ids: options.prepend ? [...addedIds, ...state.ids] : [...state.ids, ...addedIds]
     };
   }